refactor: divide fn to modules
